+++ src/Domain/Settings/Controllers/StoreStorageCredentialsController.php
@@ -2,9 +2,8 @@
 namespace Domain\Settings\Controllers;
 
 use Artisan;
+use Domain\Settings\Actions\TestFTPConnectionAction;
 use Illuminate\Http\Response;
-use Aws\S3\Exception\S3Exception;
-use League\Flysystem\UnableToWriteFile;
 use Domain\Settings\DTO\S3CredentialsData;
 use Domain\Settings\Actions\TestS3ConnectionAction;
 use Domain\Settings\Requests\StoreStorageCredentialsRequest;
@@ -12,9 +11,9 @@ use Domain\Settings\Requests\StoreStorageCredentialsRequest;
 class StoreStorageCredentialsController
 {
     public function __construct(
+        private TestFTPConnectionAction $testFTPConnection,
         private TestS3ConnectionAction $testS3Connection,
-    ) {
-    }
+    ) {}
 
     /**
      * Set new email credentials to .env file
@@ -24,20 +23,27 @@ class StoreStorageCredentialsController
         // Abort in demo mode
         abort_if(is_demo(), 204, 'Done.');
 
+        // Get storage driver from request
+        $driver = match ($request->input('storage.driver')) {
+            's3', 'storj', 'spaces', 'wasabi', 'backblaze', 'oss', 'other' => 's3',
+            'local' => 'local',
+            'ftp' => 'ftp',
+        };
+
         if (! app()->runningUnitTests()) {
-            // Test s3 credentials
-            if ($request->input('storage.driver') !== 'local') {
-                try {
-                    // connect to the s3
-                    ($this->testS3Connection)(S3CredentialsData::fromRequest($request));
-                } catch (S3Exception | UnableToWriteFile $error) {
-                    return response([
-                        'type'    => 's3-connection-error',
-                        'title'   => 'S3 Connection Error',
-                        'message' => $error->getMessage(),
-                    ], 401);
-                }
-            }
+
+            // Test driver connection
+            match ($driver) {
+                's3' => ($this->testS3Connection)(
+                    S3CredentialsData::fromRequest($request)
+                ),
+                'ftp' => ($this->testFTPConnection)([
+                    'host'     => $request->input('storage.ftp.host'),
+                    'user'     => $request->input('storage.ftp.user'),
+                    'password' => $request->input('storage.ftp.password'),
+                ]),
+                default => null
+            };
 
             $drivers = [
                 'local' => [
@@ -45,16 +51,19 @@ class StoreStorageCredentialsController
                 ],
                 's3'    => [
                     'FILESYSTEM_DISK'      => 's3',
-                    'S3_ACCESS_KEY_ID'     => $request->input('storage.key') ?? null,
-                    'S3_SECRET_ACCESS_KEY' => $request->input('storage.secret') ?? null,
-                    'S3_DEFAULT_REGION'    => $request->input('storage.region') ?? null,
-                    'S3_BUCKET'            => $request->input('storage.bucket') ?? null,
-                    'S3_URL'               => $request->input('storage.endpoint') ?? null,
+                    'S3_ACCESS_KEY_ID'     => $request->input('storage.s3.key') ?? null,
+                    'S3_SECRET_ACCESS_KEY' => $request->input('storage.s3.secret') ?? null,
+                    'S3_DEFAULT_REGION'    => $request->input('storage.s3.region') ?? null,
+                    'S3_BUCKET'            => $request->input('storage.s3.bucket') ?? null,
+                    'S3_URL'               => $request->input('storage.s3.endpoint') ?? null,
+                ],
+                'ftp' => [
+                    'FILESYSTEM_DISK' => 'ftp',
+                    'FTP_HOST'        => $request->input('storage.ftp.host') ?? null,
+                    'FTP_USERNAME'    => $request->input('storage.ftp.user') ?? null,
+                    'FTP_PASSWORD'    => $request->input('storage.ftp.password') ?? null,
                 ],
             ];
-
-            // Get storage driver from request
-            $driver = 'local' === $request->input('storage.driver') ? 'local' : 's3';
 
             // Storage credentials for storage
             setEnvironmentValue($drivers[$driver]);